Stop truncating island serials to 16 bits in the shape loop
Widening island serials to 22 bits missed three narrowing conversions in
collect_events. Serials now exceed 65535, so those truncations corrupt them --
a bug the widening itself introduced, since at 16 bits the narrowing was a
no-op.

The damaging one is promo_event_index. It is keyed by serial, and the shape
loop looked up a truncated value, so for any island whose serial passed 65535
the lookup missed and no chunk ids were attached to its promotion. The client
was told an island exists and given no members for it, so nothing bound to it
and nothing rendered -- while the chunks stayed present in physics and could
still be shot.

That matches the reported reproduction exactly: cut a building clean in half
and the top half, which is one large promotion, goes invisible; shoot where it
should be and pieces appear as they separate into further islands. It is also
why the fault grew as a session went on, since serials only cross 65535 after
sustained destruction.

The same truncation corrupted migration.to_island / from_island and the
node_to_serial cache, so chunks that did migrate were pointed at wrong islands.

Measured on the heavy-destruction spec, which reproduces this without a
browser:

  chunks drawn below ground   1419 -> 281
  lowest chunk               -4506 m -> -130 m

Better by a lot, and not fixed: 281 chunks still land below the world. The
remaining fault is separate from the truncation and still open. Notably the
magnitude is now bounded in a way that suggests a genuine offset error rather
than corrupted identity, which is a different search than the one just closed.

+++ physx-bridge/src/destruction.cc
@@ -570,7 +570,11 @@ void DestructionManager::collect_events(Slot &slot) {
   std::unordered_map<ExtStressPhysXId, const ExtStressPhysXBodySnapshot *>
       body_by_id;
   std::unordered_set<ExtStressPhysXId> live_bodies;
-  std::unordered_map<std::uint16_t, std::size_t> promo_event_index;
+  // Keyed by island serial, which is 22 bits. Narrowing this to uint16 silently
+  // fails the lookup below for any serial past 65535, which means a promoted
+  // island is announced to clients with no chunks attached -- the island exists
+  // but nothing is bound to it, so it renders as nothing at all.
+  std::unordered_map<std::uint32_t, std::size_t> promo_event_index;
   for (std::uint32_t i = 0; i < body_count; ++i) {
     body_by_id[bodies[i].bodyId] = &bodies[i];
     live_bodies.insert(bodies[i].bodyId);
@@ -634,11 +638,11 @@ void DestructionManager::collect_events(Slot &slot) {
     const ExtStressPhysXId old_body =
         old_body_it != previous_node_to_body.end() ? old_body_it->second : 0;
     auto old_serial_it = previous_node_to_serial.find(node);
-    const std::uint16_t old_serial =
+    const std::uint32_t old_serial =
         old_serial_it != previous_node_to_serial.end() ? old_serial_it->second
                                                        : 0;
     auto new_serial_it = slot.body_to_serial.find(new_body);
-    const std::uint16_t new_serial =
+    const std::uint32_t new_serial =
         new_serial_it != slot.body_to_serial.end() ? new_serial_it->second : 0;
 
     if (old_body != 0 && old_body != new_body) {
